mutation/mutation_compactor: validate range tombstone change before it is moved
e2c9cdb576c2186719ce7181dd6619f1994e262e moved the validation of the
range tombstone change to the place where it is actually consumed, so we
don't attempt to pass purged or discarded range tombstones to the
validator. In doing so however, the validate pass was moved after the
consume call, which moves the range tombstone change, the validator
having been passed a moved-from range tombstone. Fix this by moving he
validation to before the consume call.

Refs: #12575