view_update_generator: add get_endpoints_for_view_update()
Add a function which lists all of the endpoints to which
a view update should be sent. In the usual case there's
only one endpoint - the paired view replica, but in case
of an ongoing topology change it might also be necessary
to send the update to the pending nodes.

The impementation is based on the code present in:
view_update_generator::mutate_MV()

In order to calculate the endpoints we need access
to storage_proxy, which is available as a fields in view_update_generator,
so it was the easiest to add this function as a method of view_update_generator.

This function is only using view update generation,
so I think it's clean enough.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>