migration manager: if raft is enables sync with group0 leader before pulling a schema which is not available locally
Schema pull may fail because the pull does not contain everything that
is needed to instantiate a schema pointer. For instance it does not
contain a keyspace. This patch changes the code to issue raft read
barrier before the pull which will guaranty that the keyspace is created
before the actual schema pull is performed.

Refs: #3760
Fixes: #13211