view_update_generator: add get_storage_proxy()
During view generation we would like to be able
to access information about the current state
of view update backlogs, but this information
is kept inside storage_proxy.

A reference to storage_proxy is kept inside view_update_generator,
so the easiest way to get access to it from the view update code
is by adding a public getter there.

There's already a similar getter for replica::database: get_db(),
so it's in line with the rest of the code.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>