storage_proxy: fix duration casting in calculate_delay()
The function calculate_delay() is used to calculate the amount
of time by which client requests should be throttled to prevent overload.

The function calculates the delay based on an equation, and then compares
the result with a maximum allowed value, which is equal to the amount of
time left until timeout.

The comparison mentions that the budget is in milliseconds, and thus
casting microseconds to milliseconds could overflow, and because of
that there's some special code to handle this.

This looks wrong, the budget is actually in nanoseconds, not milliseconds,
lowres_clock::duration is an alias for std::chrono::nanoseconds.

Because of this the current logic is wrong, casting nanoseconds to microseconds
can't overflow. The code even introduced a possible overflow, because it tried
to avoid overflow using:
```
std::chrono::duration_cast<storage_proxy::clock_type::duration>(ret)
```

with storage_proxy::clock_type::duration being nanoseconds, and ret
being microseconds this can overflow.

Let's remove the faulty logic and change it to a simple cast from nano to micro.

The new code is also resistant to a possible change of storage_proxy::clock_type::duration.
If it were changed to e.g chrono::milliseconds, the std::max will fail because of type mismatch,
which will bring the developer's attention to this part of the code.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>