topology_state_machine: add check_knows_enabled_features
Adds the static `check_knows_enabled_features` function which takes a
set of locally supported features and the set of features enabled in the
cluster and verifies that the local node supports all of those features,
otherwise it throws.

The group0 state machine will soon check the features on every snapshot
and command application, and this function will be used there.